+++ lib/spreedly/transactions/transaction.rb
@@ -28,6 +28,8 @@ module Spreedly
         return RedactPaymentMethod.new(xml_doc)
       when 'RedactGateway'
         return RedactGateway.new(xml_doc)
+      when 'RecacheSensitiveData'
+        return RecacheSensitiveData.new(xml_doc)
       else
         Transaction.new(xml_doc)
       end