+++ HISTORY.md
@@ -5,6 +5,7 @@
 * Check attributes for non-nil before sending [markabe]
 * Add supports_verify option to Gateway [duff]
 * Add support for verify_on_gateway API call [duff]
+* Add support for RecacheSensitiveData transaction [duff]
 
 ## 2.0.11
 * Add support for ThirdPartyTokens [duff]